+++ src/bstone_hw_texture_mgr.cpp
@@ -87,7 +87,7 @@ namespace detail
 // XbrzTask
 //
 
-class XbrzTask :
+class XbrzTask final :
 	public MtTask
 {
 public: